Introduces proper Cursor handling
All cursors will be closed when they are not longer needed.
This also resolves the seemingly random crashes when the database file
could not be opened. This resulted from cursor beeing still openend in
one activities thread while an other activity (with its own thread)
wanted to access the database.